Build: Add TypeScript support
