+++ src/MudBlazor/MudBlazor.csproj
@@ -105,7 +105,7 @@
   <ItemGroup>
     <JsAssetInputs Include="TScripts/**/*.js" />
     <ScssAssetInputs Include="Styles/**/*.scss" />
-    <BuildScriptInputs Include="build.mjs" />
+    <BuildScriptInputs Include="build.mts" />
     <BuildScriptInputs Include="../eslint.config.ts" />
     <WebAssetInputs Include="@(JsAssetInputs);@(ScssAssetInputs);@(BuildScriptInputs);@(BunToolchainInputs)" />
     <UpToDateCheckInput Include="@(WebAssetInputs)" />
@@ -133,7 +133,7 @@
     <Exec Command="$(BunInstall)"
           WorkingDirectory="$(MSBuildProjectDirectory)" />
 
-    <Exec Command="$(BunWrapper) run build.mjs"
+    <Exec Command="$(BunWrapper) run build.mts"
           WorkingDirectory="$(MSBuildProjectDirectory)" />
   </Target>
 
@@ -155,7 +155,7 @@
     <Exec Command="$(BunInstall)"
           WorkingDirectory="$(MSBuildProjectDirectory)" />
 
-    <Exec Command="$(BunWrapper) run build.mjs"
+    <Exec Command="$(BunWrapper) run build.mts"
           WorkingDirectory="$(MSBuildProjectDirectory)" />
   </Target>
 
@@ -197,7 +197,7 @@
   
   <!--Dont Include in build output-->
   <ItemGroup>
-    <Content Remove="build.mjs" />
+    <Content Remove="build.mts" />
     <Content Remove="wwwroot/DoNotRemove.txt" />
     <Content Remove="wwwroot/css/MudBlazor.css" />
     <!--macOS hidden file (causes problems with dotnet pack)-->
