+++ src/MudBlazor/TScripts/entrypoint.ts
