+++ src/MudBlazor/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../tsconfig.json",
+  "include": [
+    "build.mts",
+    "TScripts/**/*"
+  ]
+}
