+++ src/package.json
@@ -1,7 +1,10 @@
 {
   "dependencies": {
+    "@types/bun": "^1.4.0",
+    "@types/node": "^26.2.0",
     "eslint": "10.8.1",
-    "sass": "1.103.0"
+    "sass": "1.103.0",
+    "typescript": "^7.0.2"
   },
   "trustedDependencies": ["sass"]
 }
